Add a new argument `additional-info' to `twittering-http-{get,put}'.
* twittering-mode.el: Add a new argument `additional-info' to
`twittering-http-{get,put}'. It can be used to send information to
generated processes.
(twittering-http-get): replace the argument `noninteractive' with
a new argument `additional-info', which will be sent to
`twittering-send-http-request'.
(twittering-http-post): add a new argument `additional-info',
which will be sent to `twittering-send-http-request'.
(twittering-call-api): send the information of `noninteractive' as
a member of `additional-info' to `twittering-http-get'.