Embed expressions for fontification in the formatter function.
* twittering-mode.el: Embed expressions for fontification in the
formatter function instead of calling them as a function.
(twittering-make-fontified-tweet-text): return an expression that
fontifies a given string. The regular expressions of strings being
fontified are concatenated on generating a formatter function.
(twittering-generate-format-table): embed the expression
fontifying a string in the formatter function. The expression will
be byte-compiled by `twittering-update-status-format'.