Wait a moment after finding that the process is unexpectedly dead.
* twittering-mode.el: Wait a moment after finding that the process
is dead and the query has not been finished. This prevents Emacs
from canceling the query on the middle of evaluating the
sentinel.
(twittering-wait-while): new macro.
(twittering-oauth-get-token-alist): wait a moment after finding
that the process is dead and the query has not been finished. This
prevents Emacs from canceling the query on the middle of
evaluating the sentinel.
(twittering-tinyurl-get): likewise.
(twittering-verify-credentials): likewise.
(twittering-get-list-sync): likewise.