`twittering-update-status' receives a tweet-type symbol. It replaces a timeline spec representing whether the tweet being edited is a direct message or not.
* twittering-mode.el: `twittering-update-status' receives a
tweet-type symbol instead of a timeline spec. It replaces a
timeline spec representing whether the tweet beingedited is a
direct message or not.
(twittering-update-status-function): update docstring.
(twittering-edit-setup-help): receive a tweet-type symbol as the
second argument instead of a timeline spec.
(twittering-update-status-from-pop-up-buffer): receive a
tweet-type symbol as the fourth argument instead of a timeline
spec.
(twittering-update-status-from-minibuffer): likewise.
(twittering-update-status): likewise.
(twittering-direct-message): invoke `twittering-update-status'
with a tweet-type symbol `direct-message' instead of a timeline
spec.
(twittering-enter): invoke `twittering-update-status' with a
tweet-type symbol instead of a timeline spec.