Control an attribute of a tweet being edited with an alist.
* twittering-mode.el: Control an attribute of a tweet being edited
with an alist bound to `twittering-edit-mode-info'.
(twittering-reply-recipient): removed.
(twittering-edit-mode-info): new variable, which is an alist
specifying the attribute of the tweet being edited.
(twittering-update-status-from-pop-up-buffer): control the
attribute of the tweet being edited with the new variable
`twittering-edit-mode-info'.
(twittering-edit-post-status): determine the attribute of the
tweet being posted by the variable `twittering-edit-mode-info'.