Warn that a message without mentions will be sent as a normal tweet.
* twittering-mode.el: Confirm that the edited message should be
sent as a normal tweet when it has been edited as a reply but it
does not include a mention to the cited user.
(twittering-edit-post-status): confirm that the edited message
should be sent as a normal tweet when it has been edited as a
reply but it does not include a mention to the cited user.