Render an ampersand as it is posted. An ampersand "&" in a posted tweet is encoded to "&amp;" in the corresponding JSON object. The current version of twittering-mode decodes it to "&" on rendering the tweet.
* twittering-mode.el: Render an ampersand as it is posted. An
ampersand "&" in a posted tweet is encoded to "&amp;" in the
corresponding JSON object. The current version of twittering-mode
decodes it to "&" on rendering the tweet.
(twittering-make-gap-list): support a gap generated by an
ampersand in a text.
(twittering-decode-html-entities): decode a character entity
reference for the ampersand "&".
(twittering-extract-common-element-from-json): decode character
entity references in a text of the given tweet.