Fix the arguments of curl for the versions later than 7.55.0.
* twittering-mode.el: Fix the arguments of curl for the versions
later than 7.55.0. Unless giving the arguments "--output -", there
is possibility that those versions output data to the standard
output as a text. Downloading an image may be failed due to such a
text output. To avoid such failure, the arguments "--output -" are
added by this commit.
(twittering-send-http-request-curl): add "--output" and "-" as
arguments of curl.